Fixes unnecessary casts to int

+++ upnp/src/genlib/net/sock.c
@@ -176,7 +176,7 @@ static int sock_read_write(
 	time_t start_time = time(NULL);
 	SOCKET sockfd = info->socket;
 	long bytes_sent = 0;
-	size_t byte_left = (size_t)0;
+	size_t byte_left = 0;
 	ssize_t num_written;
 
 	FD_ZERO(&readSet);
@@ -226,10 +226,8 @@ static int sock_read_write(
 			} else {
 #endif
 				/* read data. */
-				numBytes = (long)recv(sockfd,
-					buffer,
-					(int)bufsize,
-					MSG_NOSIGNAL);
+				numBytes = (long)recv(
+					sockfd, buffer, bufsize, MSG_NOSIGNAL);
 #ifdef UPNP_ENABLE_OPEN_SSL
 			}
 #endif
@@ -247,7 +245,7 @@ static int sock_read_write(
 					/* write data. */
 					num_written = send(sockfd,
 						buffer + bytes_sent,
-						(int)byte_left,
+						byte_left,
 						MSG_DONTROUTE | MSG_NOSIGNAL);
 #ifdef UPNP_ENABLE_OPEN_SSL
 				}
